optimize: unlink (delete) preview file on close request         fix(utils/temp): return if `path` parameter isn't a instance of `str` or `bytes`

+++ halftone/backend/utils/temp.py
@@ -28,5 +28,8 @@ class HalftoneTempFile:
         Delete temporary file located at given path.
         """
 
+        if not isinstance(path, (str, bytes)):
+            return
+
         temp_file = Path(path)
         temp_file.unlink()
